% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCPhosApp-package.R
\docType{package}
\name{CCPhosApp-package}
\alias{CCPhosApp}
\alias{CCPhosApp-package}
\title{CCPhosApp: A Shiny App for CCPhos}
\description{
Shiny app to facilitate working with dsCCPhosClient
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ccphos.netlify.app/}
  \item \url{https://github.com/BastianReiter/CCPhosApp}
}

}
\author{
\strong{Maintainer}: Bastian Reiter \email{b.reiter@med.uni-frankfurt.de} (\href{https://orcid.org/0000-0002-1088-0502}{ORCID})

}
\keyword{internal}
