% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{ColorToRGBCSS}
\alias{ColorToRGBCSS}
\title{ColorToRGBCSS}
\usage{
ColorToRGBCSS(Color, Alpha = 1, RenderNATransparent = TRUE)
}
\arguments{
\item{Color}{\code{character} - Vector of hexadecimal color code}

\item{Alpha}{\code{double} - Optional alpha value vector - Default: 1}

\item{RenderNATransparent}{\code{logical} - Indicating whether \code{NA} values for \code{Color} result in a totally transparent color}
}
\description{
Turn hexadecimal color code into a string of CSS code of the form 'rgba(r, g, b, a)'.
Can be used with vectors.
}
\author{
Bastian Reiter
}
