% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{ConvertLogicalToIcon}
\alias{ConvertLogicalToIcon}
\title{ConvertLogicalToIcon}
\usage{
ConvertLogicalToIcon(DataFrame)
}
\arguments{
\item{DataFrame}{\code{data.frame} or \code{tibble}}
}
\value{
\code{data.frame}
}
\description{
ConvertLogicalToIcon
}
\author{
Bastian Reiter
}
