% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTable.Counter.NestedView.R
\name{CreateTable.Counter.NestedView}
\alias{CreateTable.Counter.NestedView}
\title{CreateTable.Counter.NestedView}
\usage{
CreateTable.Counter.NestedView(CounterData)
}
\arguments{
\item{CounterData}{\code{list} - Contains...}
}
\value{
Table object of \code{reactable} class
}
\description{
Compile table check data from \code{dsFredaClient::ds.CheckTable()} or \code{dsFredaClient::ds.CheckDataSet()} into a table suited for display
}
\author{
Bastian Reiter
}
