% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTable.TableCheck.R
\name{CreateTable.TableCheck}
\alias{CreateTable.TableCheck}
\title{CreateTable.TableCheck}
\usage{
CreateTable.TableCheck(TableData)
}
\arguments{
\item{TableData}{\code{list} - Contains...}
}
\value{
\code{list} - Contains two \code{tibbles}:
\itemize{ \item 'TableDetails' - Contains coherent table check data
\item 'HeaderColspans' }
}
\description{
Compile table check data from \code{dsFredaClient::ds.CheckTable()} or \code{dsFredaClient::ds.CheckDataSet()} into a table suited for display
}
\author{
Bastian Reiter
}
