% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTableMonitor.R
\name{CreateTableMonitor}
\alias{CreateTableMonitor}
\title{CreateTableMonitor}
\usage{
CreateTableMonitor(TableData)
}
\arguments{
\item{TableData}{\code{list} - Contains...}
}
\value{
\code{list} - Contains two \code{tibbles}:
\itemize{ \item 'TableDetails' - Contains coherent table check data
\item 'HeaderColspans' }
}
\description{
Compile table check data from \code{dsCCPhosClient::ds.CheckTable()} or \code{dsCCPhosClient::ds.CheckDataSet()} into a table suited for display
}
\author{
Bastian Reiter
}
