% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameToHtmlTable.R
\name{DataFrameToHtmlTable}
\alias{DataFrameToHtmlTable}
\title{DataFrameToHtmlTable}
\usage{
DataFrameToHtmlTable(
  DataFrame,
  TableID = NULL,
  CategoryColumn = NULL,
  ColContentHorizontalAlign = "left",
  ColumnCSSClass = NULL,
  ColumnIcons = NULL,
  ColumnLabels = NULL,
  ColumnLabelsLineBreak = FALSE,
  ColumnMaxWidth = 50,
  HeaderColspans = NULL,
  RotatedHeaderNames = character(),
  RowColorColumn = NULL,
  SemanticTableCSSClass = "ui celled table",
  TableStyle = "",
  TurnColorValuesIntoDots = FALSE,
  TurnLogicalsIntoIcons = FALSE,
  TurnNAsIntoBlanks = FALSE
)
}
\arguments{
\item{DataFrame}{\code{data.frame} or \code{tibble}}

\item{TableID}{\code{string} - Used to identify the table object in the DOM}

\item{CategoryColumn}{\code{string}}

\item{ColContentHorizontalAlign}{Either a single string for all columns or a named character vector to determine horizontal content alignment for specific columns}

\item{ColumnCSSClass}{Either a single string for all columns or a named character vector to determine table cell classes for specific columns}

\item{ColumnIcons}{named \code{vector}}

\item{ColumnLabels}{named \code{vector} with names being original table column names and elements being displayed column labels}

\item{ColumnLabelsLineBreak}{\code{logical} Indicating whether column labels should span across two lines}

\item{ColumnMaxWidth}{\code{integer} - Maximum width of columns in character spaces}

\item{HeaderColspans}{\code{integer vector}}

\item{RotatedHeaderNames}{\code{vector}}

\item{RowColorColumn}{\code{string}}

\item{SemanticTableCSSClass}{\code{string}}

\item{TableStyle}{A \code{string} containing CSS style elements}

\item{TurnColorValuesIntoDots}{\code{logical}}

\item{TurnLogicalsIntoIcons}{\code{logical}}

\item{TurnNAsIntoBlanks}{\code{logical}}
}
\value{
HTML code
}
\description{
Turn a data.frame into HTML table code.
}
\details{
To enable cell-specific CSS classes or style code, the passed data.frame should have columns named 'CellCSSClass.ABC' where 'ABC' is another column in the data.frame.
}
\author{
Bastian Reiter
}
