% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FredaGUI-package.R
\docType{package}
\name{FredaGUI-package}
\alias{FredaGUI}
\alias{FredaGUI-package}
\title{FredaGUI: A Graphical User Interface for FREDA}
\description{
Shiny app to facilitate working with dsFredaClient
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ccphos.netlify.app/}
  \item \url{https://github.com/BastianReiter/FredaGUI}
}

}
\author{
\strong{Maintainer}: Bastian Reiter \email{b.reiter@med.uni-frankfurt.de} (\href{https://orcid.org/0000-0002-1088-0502}{ORCID})

}
\keyword{internal}
