% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartCCPhosApp.R
\name{StartCCPhosApp}
\alias{StartCCPhosApp}
\title{StartCCPhosApp}
\usage{
StartCCPhosApp(
  ADSTableCheckData = NULL,
  CCPTestData = NULL,
  CDSTableCheckData = NULL,
  CurationReportData = NULL,
  DSConnections = NULL,
  RDSTableCheckData = NULL,
  ServerSpecifications = NULL,
  ServerWorkspaceInfo = NULL,
  EndProcessWhenClosingApp = TRUE,
  RunAutonomously = TRUE,
  RunInViewer = FALSE,
  ...
)
}
\arguments{
\item{CCPTestData}{\code{list} - Optional CCP test data}

\item{CurationReportData}{\code{list} - Optional CurationReport data}

\item{DSConnections}{\code{list} of \code{DSConnection} objects}

\item{RDSTableCheckData}{\code{list} - Optional RDSTableCheck data}

\item{ServerSpecifications}{\code{data.frame}}
}
\description{
Launch Shiny app
}
\author{
Bastian Reiter
}
