% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartFredaApp.R
\name{StartFredaApp}
\alias{StartFredaApp}
\title{StartFredaApp}
\usage{
StartFredaApp(
  TestData = NULL,
  RDSCheckData = NULL,
  CDSCheckData = NULL,
  ADSCheckData = NULL,
  CurationReport = NULL,
  DSConnections = NULL,
  ServerSpecifications = NULL,
  ServerWorkspaceInfo = NULL,
  EndProcessWhenClosingApp = TRUE,
  RunAutonomously = FALSE,
  RunInViewer = FALSE,
  ...
)
}
\arguments{
\item{TestData}{\code{list} - Optional test data}

\item{RDSCheckData}{\code{list} - Optional RDSCheck data}

\item{CurationReport}{\code{list} - Optional CurationReport data}

\item{DSConnections}{\code{list} of \code{DSConnection} objects}

\item{ServerSpecifications}{\code{data.frame}}
}
\description{
Launch Shiny app
}
\author{
Bastian Reiter
}
