% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget.ProcessingMonitor.R
\name{Widget.ProcessingMonitor}
\alias{Widget.ProcessingMonitor}
\title{Widget.ProcessingMonitor}
\usage{
Widget.ProcessingMonitor(
  ServerSpecifications = NULL,
  DSConnections = NULL,
  EnableLiveConnection = TRUE,
  EndProcessWhenClosingApp = TRUE,
  RunAutonomously = TRUE,
  RunInViewer = FALSE,
  UseVirtualConnections = FALSE,
  ...
)
}
\arguments{
\item{ServerSpecifications}{\code{data.frame} containing credentials for login}

\item{DSConnections}{\code{list} of \code{DSConnection} objects}

\item{EndProcessWhenClosingApp}{\code{logical} indicating whether the background process that runs the Shiny app (if it runs autonomously) should end when the app is closed (default) or should be preserved, in which case it should be ended manually.}

\item{RunAutonomously}{\code{logical} indicating whether the Shiny app is hosted by a background process (default) available as a URL via web browsers or - if set to \code{FALSE} - is hosted by the current running R session.}

\item{RunInViewer}{\code{logical} indicating whether the Shiny app should be run in the RStudio Viewer pane (Default: \code{FALSE})}
}
\value{
When 'RunAutonomously' is set to \code{TRUE} this function can return the background process to make it assignable to an R symbol. Otherwise it will run/return a \code{shinyApp} object.
}
\description{
Launch a Shiny app that facilitates interacting with output of processing monitoring, like data transformation tracks and data set checks.
}
\author{
Bastian Reiter
}
